@@ -1,11 +1,14 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
-import App from "./App";
+import React from "react";
+import ReactDOM from "react-dom/client";
 import "./index.css";
+import App from "./App";
+import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
-  
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
     <App />
- ,
-  document.getElementById("root")
+  </React.StrictMode>
 );
+
+reportWebVitals();
